Fix color render bug
Deleted getRandomColor function - this function was ran each time the shoe component's state changed and therefore re-rendered.  To get around this, I simply added a color field for each shoe into the data and database, then rendered the number of colors once the data was mapped over in the parent component.